backend/requirements.txt
Flask
flask_sqlalchemy
python-dotenv
Flask-Cors
